+++ mayan/apps/common/tests/mixins.py
@@ -4,8 +4,16 @@ import glob
 import os
 
 from django.conf import settings
+from django.contrib.auth import get_user_model
+from django.contrib.auth.models import Group
 if getattr(settings, 'COMMON_TEST_FILE_HANDLES', False):
     import psutil
+from permissions.models import Role
+from permissions.tests.literals import TEST_ROLE_LABEL
+from user_management.tests import (
+    TEST_ADMIN_PASSWORD, TEST_ADMIN_USERNAME, TEST_ADMIN_EMAIL,
+    TEST_GROUP_NAME, TEST_USER_EMAIL, TEST_USER_USERNAME, TEST_USER_PASSWORD
+)
 
 from ..settings import setting_temporary_directory
 
@@ -103,3 +111,22 @@ class TempfileCheckMixin(object):
                 )
             )
         super(TempfileCheckMixin, self).tearDown()
+
+
+class UserMixin(object):
+    def setUp(self):
+        super(UserMixin, self).setUp()
+        self.admin_user = get_user_model().objects.create_superuser(
+            username=TEST_ADMIN_USERNAME, email=TEST_ADMIN_EMAIL,
+            password=TEST_ADMIN_PASSWORD
+        )
+
+        self.user = get_user_model().objects.create_user(
+            username=TEST_USER_USERNAME, email=TEST_USER_EMAIL,
+            password=TEST_USER_PASSWORD
+        )
+
+        self.group = Group.objects.create(name=TEST_GROUP_NAME)
+        self.role = Role.objects.create(label=TEST_ROLE_LABEL)
+        self.group.user_set.add(self.user)
+        self.role.groups.add(self.group)