+++ mayan/apps/document_states/admin.py
@@ -30,7 +30,7 @@ class WorkflowAdmin(admin.ModelAdmin):
 
     filter_horizontal = ('document_types',)
     inlines = (WorkflowStateInline, WorkflowTransitionInline)
-    list_display = ('label', 'document_types_list')
+    list_display = ('label', 'internal_name', 'document_types_list')
 
 
 @admin.register(WorkflowInstance)