+++ HISTORY.rst
@@ -6,10 +6,12 @@
   FUSE index mirror. GitLab issue #385
 - Add support for check for the latest released version of Mayan from the
   About menu.
+- Support for rebuilding specific indexes. GitLab issue #372.
 - Rewrite document indexing code to be faster and use less locking.
 - Use a predefined file path for the file lock.
 - Catch documents with not document version when displaying their thumbnails.
 - Document page navigation fix when using Mayan as a sub URL app.
+- Add support for indexing on workflow state changes.
 
 2.2 (2017-04-26)
 ================